ordered all profiles
Added the `ordinal` function to every .station file by popularity and usefulness then alphabetically. All stations are `ordinal: 99` except for stations in the main "ZLA" file which are ordered:
- KLAX (0)
- KLAS (1)
- KSAN (2)
- KSNA (3)
- KBUR (4)
- KONT (5)
- KVNY(6)
- KPSP (7)
- KSBA (8)

This has the downstream effect of pinning the "popular" airport to the top of other profiles in addition to the main profile.