Added Ordinals, Added BFL, Updated Readme (#13)
* Ordered all profiles

Added the `ordinal` function to every .station file by popularity and usefulness then alphabetically. All stations are `ordinal: 99` except for stations in the main "ZLA" file which are ordered:
- KLAX (0)
- KLAS (1)
- KSAN (2)
- KSNA (3)
- KBUR (4)
- KONT (5)
- KVNY (6)
- KPSP (7)
- KSBA (8)
- KBFL (9)

This has the downstream effect of pinning the "popular" airport to the top of other profiles in addition to the main profile.

* Created development compiler

Added "-DEV" to file names in compiler.py to assist in A/B testing changes in vATIS

* Added BFL and Ordinals to IFP, GCN

Bakersfield was missing from the primary ZLA profiles. Added it to both `ZLA` and `ZLA 2` with ordinal 99, added ordinal 99 to GCN and IFP.

* Sorted airports by "popularity," added BFL

* Synced DEV to ordering branch

Added instructions for compiler and devcompiler. Defined new files.

* Renamed DEV files to sort DEV ahead of facility

* Grammar and formatting update to readme